Schedule timeouts for API scans
Tests in next CL.

BUG=31803601

+++ OpenScienceJournal/whistlepunk_library/src/main/java/com/google/android/apps/forscience/javalib/Delay.java
@@ -26,7 +26,7 @@ public class Delay {
         return new Delay(secs, TimeUnit.SECONDS);
     }
 
-    public static Delay millis(int millis) {
+    public static Delay millis(long millis) {
         return new Delay(millis, TimeUnit.MILLISECONDS);
     }
